Move GPDB-specific GetTupleVisibilitySummary functions to separate file.
tqual.c is quite a long file, and there are plenty of legitimate
MPP-specific changes in it compared to upstream. Move these functions to
separate file, to reduce our diff footprint. This hopefully makes merging
easier in the future.

Run pgindent over the new file, and do some manual tidying up. Also, use
CStringGetTextDatum() instead of the more complicated dance with
DirectFunctionCall and textin(), to convert C strings into text Datums.
