+++ src/backend/utils/time/Makefile
@@ -14,6 +14,6 @@ include $(top_builddir)/src/Makefile.global
 
 OBJS = combocid.o tqual.o
 
-OBJS += sharedsnapshot.o
+OBJS += sharedsnapshot.o visibility_summary.o
 
 include $(top_srcdir)/src/backend/common.mk
